backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
websockets
pydantic
pyasn1-modules